+++ src/core/model/layers/layer.hpp
@@ -19,7 +19,7 @@ class Layer : public AnimationContainer
     GLAXNIMATE_PROPERTY_REFERENCE(Layer, parent, &Layer::valid_parents, &Layer::is_valid_parent)
     GLAXNIMATE_PROPERTY(float, start_time, 0, {}, {}, PropertyTraits::Visual)
     GLAXNIMATE_SUBOBJECT(Transform, transform)
-    GLAXNIMATE_ANIMATABLE(float, opacity, 1, {}, 0, 1)
+    GLAXNIMATE_ANIMATABLE(float, opacity, 1, &Layer::opacity_changed, 0, 1)
 
 public:
     class ChildLayerIterator
@@ -117,6 +117,9 @@ protected:
 
 private slots:
     void on_transform_matrix_changed();
+
+signals:
+    void opacity_changed(float op);
 
 private:
     Composition* composition_;