+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     * Fixed palettes not being saved correctly
     * Undoing commands no longer adds spurious keyframes
     * Dock widgets have their icon displayed correctly on all supported systems
+    * Opacity is now being displayed correctly
 
 ## 0.1.0